blogging_project/requirements.txt
# Django and Django REST Framework
Django>=5.1.2
djangorestframework>=3.14.0

# Additional dependencies that might be useful
# Uncomment if you need these features:
# djangorestframework-simplejwt>=5.2.0  # For JWT authentication
# django-cors-headers>=4.0.0  # For CORS support
# python-decouple>=3.8  # For environment variable management
# psycopg2-binary>=2.9.0  # For PostgreSQL support (if not using SQLite)

# Development dependencies (optional)
# pytest>=7.0.0  # For testing
# pytest-django>=4.5.0  # Django test integration
# coverage>=7.0.0  # For code coverage
# black>=23.0.0  # For code formatting
# flake8>=6.0.0  # For linting
